Work around PacBio HiFi quality induced biases.
With a homopoymer, eg in ATTTTC, one base in the run is likely given a
low confidence value.  This is always the same end.  After reverse
complementing during alignment, we get a mix of low qualities at the
first and last base.  However the alignment algorithms typically do
left-justification of gaps, so any slight strand bias will be
exacerbated as high vs low confidence calls for that indel.

This method averages the first and last qualities together to get an
amortised new quality.  It then averages the end two inner qualities
together, and so on.

This has been seen to reduce indel consensus call errors considerably
in some PB HiFi data sets.