Also change the new pileup function quality scoring for gaps.
Gaps in alignments don't have confidence values as they don't have a
base assigned to them.  We previously used the average of the two
neighbouring bases, which works OK for Illumina given a fairly even
profile of quality values.

With wildly varying qualities in homopolymers however, the overcall /
undercall is much better correlated to the minimum of the neighbouring
bases rather than the average.

Note that mathematically this is much closer to the real thing anyway,
as averaging the error probability for very-low and very-high should
really turn from phred to prob, average, and turn back to phred.  In
the limit with extreme variance in phred scores, that ends up at low+3.
So minimum is a good enough proxy too.