Allow ampliconstats bed file to use fewer refs than the input file
Fixes a bug where references mentioned in the input file headers
but not in the bed file would cause it to complain that the
SAM headers were not consistent.  It is actually reasonable to
do this, as when genotyping large genomes PCR products may only
be made on a subset of the chromosomes.

The test used to skip the unused refs (!amps[r].sites) is already
used in other parts of the code for the same purpose.