Stop reads wrongly being removed in hard clip mode. (PR #1722)
The pre hard clip check was wrongly removing entire reads.  The large clip was bigger than the query length (calculated with the CIGAR string) but not the length against the reference (which is what was needed).

The new version of the test needs to detect if it's run out of reference bases.  This is most easily
done by moving the code handling this case to the part that deals with the CIGAR walk reaching
the end if the CIGAR data - where this happens, no aligned bases will be left.