Revise collate to require an input filename.
As per discussion in #1724, we now require an explicit "-" input
filename.

This also fixes the bug where "samtools collate" without any argument
writes the usage to stderr instead of stdout.

Note the logic changed previously in the bamshuf function is still
there, as this is a harmless additional check incase we ever reverse
the main argument parsing.  For now that ?: operator will always be true.