Add consensus --mark-ins option.
This adds underscore before inserted bases and their quality values.

The reason for this is that with "-a --show-del yes" we can now
determine a mapping between consensus coordinates and their original
reference coordinates, in turn facilitating easy alignment of the two
to record differences.

This differs slightly to bcftools consensus which uses upper/lowercase
to mark indels.  That works there,  but bcftools cannot report
heterozygous indels.  We use the case for that purpose (with "A" being
A/A and "a" being A/*).