Fix samtools consensus buffer overrun with MD:Z handling.
An MD tag that refers to bases beyond the SEQ field could
write beyond the end of the local_nm buffer.  We already had bounds
checks in 2 out of 3 initialisation loops, but sadly missed the first
one.  The code adds +5 to an array entry, rather than arbitrary
user-specified data, but it's still a security risk to this one
sub-command.

Also streamlined the process to simply skip data with qlen of zero
(SEQ "*").

Fixes #1744