Fix a buffer read-overflow on sequences with seq "*".
When a record has no seq entry, but still has a CIGAR string, tview
attempts to still display the sequence.  Trying to hide it is
challenging as it needs changes in the lpileup code too, and
potentially there is still merit in showing the alignment minus bases
if that is indeed what we recorded.

However the sequence and qualities show are fetched from beyond the
end of the seq and qual pointers, so it's a read buffer overrun into
arbitrary memory.  They've now been replaced with N and qual 0.

Note htslib has a similar related bug.

It's amazing we've never spotted this before!