Improve samtools cram_size subcommand.
The default output is similar to Io_lib's cram_size program.
However we also have a verbose mode that separates blocks by type of
compression used, plus a -e (--encodings) option to report the
encoding methods used per container.  This is similar to (part of)
cram_dump, but limited in scope.

Another difference to io_lib's tool is that we deliberately permit
output on files containing errors, so we can report a break down so
far of a file that is in the process of being written still.  This
will return a non-zero exit code however to acknowledge the inevitable
decoding errors.