Expand fasta/fastq man page
This is just a minor update to explain the rationale - to recreate the
input FASTQ meaning one record per primary alignment and not one
record per SAM alignment irrespective of secondaries /
supplementaries.

Also document that the sequence is as per SAM SEQ field, so includes
soft-clips and (somewhat obviously) doesn't include hard-clips!

Fixes #1792