Minor fix to wording of mpileup --rf.
The change of "include reads" to "only include reads" is stronger
language, emphasising that this is a filter-out rather than filter-in
rule.  The use of "only" already appears in view, depth, fastq and
consensus subcommands.

I note however that "coverage" currently says

    ... required flags: skip reads with mask bits unset []

This was my initial thought on phrasing this, as it's explicitly what
the code does. However I chose to go with the majority for a
smaller change instead. (Arguably we should also rephrase coverage.)

Also fixed line-wrapping to appear better on a standard 80 column
terminal and to get the usage statement with consistent indentation.

Added text to the man page to make explicit that --rf and --ff work
independently and don't override each other.

Fixes #1791