Add minimiser sort option to collate by a fasta (indexed).
Just as we compute minimisers per read, we can also index a known
reference fasta to produce a minimiser per window.

By default our minimiser strategy is simply collating a set of reads
together that share the same minimiser, sub-ordered by the position of
that minimiser within each read.  However from one minimiser value to
the next the ordering is arbitrary.  Instead our ordering could be
based on the order of those minimisers occurring within the reference
genome.

This is basically an extremely crude and simplistic read aligner, but
it's exploiting the same tricks for boosting read compression so we
don't need the alignment to be correct in every case.

To further improve results, we don't just compute the smallest
minimiser for the entire read, but instead the smallest which has a
unique match in the indexed reference.  This can help reduce the
impact of errors / variants.

TODO: consider something better than minimiser though, eg syncmer.
However that's a different topic.

Some benchmarks on 50x HiFi coverage of D. melanogaster (SRR10238607):
BA size, when compressed with zstd (*).

Original fastq order:   6364702404  24.63%
sort -M:                 714933310   2.77%
sort -MI:                508739466   1.97%
sort -MR:                685390622   2.65%

CPU wise it's about 30% slower than the non-indexed minimiser,
including the time to index the genome.  However this increases to
150% slower for short read data, as we have many more hashes to lookup.

(*) zstd isn't a standard CRAM codec, but using this as it's a much
faster proxy for lzma which is slow to test on.