Fix a sort -M bug when merging sub-blocks.
Introduced in 038cde8 we reuse bam_merge_simple for both internal
merging before writing out to temporary files and the final merging.

This new first layer of merging meant that our reset of the temporary
fields used to house the unmapped minimiser hash were reset too early,
causing the subsequent final merge to perform poorly.  The data was,
but in a suboptimal ordering for compression.