Adds homopolymer squashing capacity to sort -M
This collapses runs of the same base type together when computing the
hash.  Eg GAATCCAAATG becomes GATCATG.  This is a useful
transformation for instruments with a higher homopolymer indel error
rate.

An example of BA compressed size (*) with Revio:

          Without -H            With -H
-M        73571039 46.0s        63952159 63.2s
-MI       51562640 92.0s        49747975 87.3s

(*) Using an experimental CRAM with zstd support, but the above is
also a good size proxy for the slower lzma algorithm.

Homopolymer collapsing does come with a CPU cost, but if we're
optimising for size it's possibly worth it.