Make calmd "work" on unaligned data or empty files.
If a processing pipeline (eg pooling, barcode splitting, etc) has a
chance to create a file with headers but no sequences, then it should
not be an error.  It can instead simply repeat the header and exit
cleanly.

Similarly a SAM file containing only unaligned data will have no
alignments to rebuild the MD tag for, which also should not be an
error.

We do however still produce a warning, as it's an unlikely scenario
and it may be symptomatic of something having gone wrong earlier in
processing.

Fixes #1839