Extend import --order TAG to --order TAG:length
If we specify a tag length, it implicitly changes the tag format from
integer to string and uses a 0-padded number (in string form) as the
tag.  This works around issues of more than 4 billion records, where
BAM and CRAM cannot encode the tag data.

We could use an alternative encoding, such as base64, but it looks
like in CRAM 3.1 it's around 5% smaller tag (rans4x16pr -o65), which
as a percentage of the entire file is tiny.  If we use bzip2 that gap
reduces to ~3%.

Fixes #1847