Add a "samtools consensus -aa" mode. (#1851)
* Add a "samtools consensus -aa" mode.

This is like mpileup -aa and depth -aa.  The single `a` reports all
bases in contigs covered by alignments, but does skips contigs with no
overlapping data at all.  The double `aa` (or `-a -a`) reports Ns even
for the references with no alignments against them.

Fixes #1849

* Fix "consensus -aa" with explicit regions.

We don't need to report other chromosomes when we're limiting output
to a single chromosome.  Even worse it was previously possible to get
infinite loops when specifying -aa and regions.

Also fixed a repetition in the man page,

(Both issues reported by Vasu Sarma)