Add view -N/-R ^FILE for FILE negation.
These options output only the reads with read names (-N) or read
groups (-R) listed in FILE.  Copying the syntax used in various
bcftools commands, now ^FILE negates this so it only outputs reads
with their name/group not listed in the file.

The third exclusion-by-hash in the code is tag types, but it's less
likely we'll be using files of tag names as these tend to be more
prescribed than the generated names and read-groups.  Hence negation
isn't supported there.

Fixes #1895