Fix possible array bounds violation in bed_minoff()
Confusion over how to check the number of items in
bed_reglist_t::idx could lead to either attempting to read off
the end of the array, or returning a value smaller than it should
have.  The latter was harmless, apart from making lookups slightly
slower, but the former could lead to crashes or results not being
returned.

The correct limit wasn't actually stored directly although it
could be calculated from the last range in bed_reglist_t::a.  For
easier look-ups, it's now stored in bed_reglist_t::max_idx.

bed_index_core() is rewritten to store max_idx; add more error
and input range checking; and be more efficient.  It no longer
leaves gaps (filled with -1) in the index it builds, so the code
that dealt with them in bed_minoff() can be removed.

Some additional changes are made so failures can be passed back
and handled appropriately.