Add MPC to the samtools stats man page
This is a bit convoluted when it comes to what is a match and what is
not.  For example N in seq and N in reference is added to the "N"
column.

"R" in the reference is not counted as mismatch, even if the sequence
is "C" (R is A or G). Matches are case insensitive.

See discussion in #1954