Add a CRAM file size regression test
This mainly tests htslib, but it doesn't have a good large general
purpose CRAM file for validation.  We don't really have a large enough
one here either, but this is sufficient.

We could create our own artificial one, but the current test spots the
problem in #1968 so it's an easy starting point.  Are there zlib,
bzip2, or lzma implementations which break this? I tested zlib and
libdeflate, and our check is 2% over that so this is a reasonable
buffer.

If it breaks, we can amend the testing.