Change default max depth for bedcov.
The -d option is used for specifying the minimum depth we need to meet
in order to count bases covered with at least that depth in a new
column.  It also ups the bam_mplp_set_maxcnt mpileup limit from 64000
to the '-d' value if bigger.

Therefore extremely deep data requires the use of '-d' to count bases
even when it's not interested in the additional column.

Changed the default to be INT_MAX (every read is counted) and added a
new --max-depth option to control this so we can limit it for
problematic files if desired.

Also fixed usage statement formatting.

Fixes #1950