Speed up mpileup by reducing passes through arrays and qual (#1995)
The code has multiple passes through plp[i] getting ther qualities and
comparing against conf->min_baseQ.  This is done for printing the
depth, which sequence bases to display, as well as obviously the
quality values.  All 3 of these are now computed in a single step,
putting the values into kstrings.

We still have additional excess loops if we ask for the optional
columns, but that's not the default output and can be dealt with at a
later stage.

Also remove some of the excess seq_nt16_table and seq_nt_str back and
forth, plus specialise for strand so we don't also have to call
toupper and tolower plus detection of the "=" code.

The combined speed up is 20-25% on the BAM files I tested (long and
short read).

Also adjusted the main loop to avoid variable shadowing. (That's
nothing to do with the speed up changes.)