Make name sort put primary reads before secondary reads.
For completeness, we consider non-secondary supplementaries as before
secondaries, and non-supplementary secondaries before
secondary+supplementary.

This is because a (non-2ndary) supplementary read is a continuation of
the primary read and should be grouped before alternate alignments for
this read.

We still have all READ1 data before all READ2 data, so the primary
alignment has not changed ordering.  This is simply a better way of
resolving ties than keeping the original ordering.

Fixes #2010