Fix a bug in mpileup -a and ref bases.
When specifying -f ref.fa to report the reference and using the "-a"
option, once we hit the end of one reference and before proceeding to
then ext we report all remaining positions in that reference.  However
we had already loaded the next reference sequence up, so the bases
reported were wrong.

It's been this way since 2016 when the -a option was added in b4e452c.
So I am assuming the fact it's taken 8 years to be spotted means this
is pretty rare for this to affect people (given it only changes
locations after the alignments have finished).

Fixes #2018