Fix new clang -pedantic warnings about func()
This warning appears with c90 and disappears again in c23.  It's
harmless, but we do build htslib with ultra-pedantry so applying the
same logic here.  (Although I would also agree with the camp that says
"if it hurts when you do that, then don't do that!")

Also fixed a couple other minor issues.  The code now builds with
    -D_XOPEN_SOURCE=600 -std=c99 -pedantic

Note the strnlen to strlen changes are OK as we can guarantee all uses
are using nul-terminated strings, either because we wrote them
ourselves with snprintf or because they're declared as such in the
htslib API (header "target_name" array).  Potentially it changes
behaviour for specifying 48-character prefix matches of references
with names over 48 characters long, but don't do that!  Use a sensible
reference. (If we really care we could just increase TV_MAX_GOTO with
some care to the visual sizing.)