Don't look at chr "*" for unmapped-placed reads with view --fetch-pairs
The logic previously included `rec->core.flag & BAM_FMUNMAP` but a
placed read while unmapped does have a chromesome and position.
Searching in HTS_IDX_NOCOOR is both a very expensive operation and
also won't actually find the pair we're looking for.

Also fixed the test data.  It looks like one record was manually
edited to move it to the unmapped chr "*", but none of the other
fields were corrected.  Hence it breaks the assertion in the man page
about RNEXT, PNEXT and TLEN being correct.

Fixes #2059