Make ampliconclip primer counts output deterministic
The primer-counts output was being written with references
in the order stored in the hash table, which is pseudo-randomised
by the hash function.  To get a more stable ordering, keep a
list of references in the order they appear in the bed file, and
then iterate through that.

ampliconstats uses the same bad file reading code, but already
outputs in the order of references in the bam header, so the
updated loader allows it to skip building the reference list from
the bad file.  (ampliconclip could in theory work the same way,
but as it doesn't insist that all bed references need to be in the
bam header, and outputs zero counts for the missing ones, it's
simpler to go with the bed file ordering so the output doesn't
change more than it needs to.)