Handle alignments without quality.
Secondary alignments with CIGAR strings but without quality (valid in the sam spce) could cause segfaults and/or uninitialised memory access.

This commit fixes that and changes the quality calculations to just use existing qualities.  This changes existing behaviour but as secondary values are excluded by default the effects should be minor.