Fix undefined behaviour in samtools fastq with empty QUAL.
Incrementing via "qt_end += qt != NULL" will be qt_end+=0 when we have
a null qt/qt_end, but NULL+0 is technically undefined.

    bam_fastq.c:852:30: runtime error: applying zero offset to null pointer
    SUMMARY: UndefinedBehaviorSanitizer: undefined-behavior bam_fastq.c:852:30 in

Easy to just replace with two loops and avoid such cleverness.  It's
not a major portability issue, but it's nice to have no warnings from
ubsan as it makes finding other problems easier.