Add support for compiling with -Wformat-nonliteral.
This is enabled via -Wformat=2, but it causes a number of warnings.
Some are trivial to fix, needing simply a format attribute, others
less so.

Specifically ampliconclip constructs a format string based on the
number of columns detected in the bed file.  Fundamentally this
isn't compatible with the format checking.  However Rob points out an
easier way of doing this by looking at sscanf return values, and it
turns out this also fixes a whole swathe of unrelated recent
ampliconclip bugs.

Also fixed a newly detected formatting problem (albeit 32-bit
platforms only and only in error reporting) in tmp_file.c.