Simplify ampliconclip bed parsing, plus bug fix.
The size check for "num_columns > sizeof(scanf_components)" is wrong,
meaning with more columns than 6 we had a read buffer overrun.  The
code also required tabs, but the BED format does not strictly state
this.

The original code prior to #2033 just uses sscanf and no tab counting,
which is better as it supports any white space and matches the BED
spec (bar the weird oddity of permitting spaces in names if only
single tabs are used, but that's a nasty case we have no desire to
support).

Instead we rely on the return value from sscanf to inform us of the
(minimum) number of columns detected.

Also amended the test data to incorporate up to 12 columns and down to
3 to validate the code can still parse correctly, plus spaces instead
of tabs.