Work around CFLAGS overriding --enable-werror.
If we have CFLAGS set then ./configure --enable-werror has no impact,
as it just sets this in CFLAGS rather than another variable.

In theory we should do ./configure --enable-werror CFLAGS=xyz.
However that means parameterising the <<:*COMPILE template.  It's
easiest to do this via an environment given the matrix, but we have
to be careful to avoid it undoing itself again.  We also need to take
care to not do --enable-werror when using ubsan.

This is a long-standing issue of removal of -Werror that's been
present right since the first day we used cirrus-CI.