Fix bed_index_core() handling of overlapping entries
It needs to skip if the current entry ends before the previous
one, which can happen if the input BED file has entries that
are completely contained in an earlier region.  Failing to
do this could cause last_end to jump backwards, resulting
in some incorrect entries being put in the index.  This could
lead to hits between the end of the innermost and outermost
nested BED entries being missed.

Fixes #2104 ("samtools view -L " incomplete output for BED with
nested targets)