Remove -e from .cirrus.yml make invocations
"make -e" causes variables to be set from the environment,
overriding those from configure, which is not what we want.

Where set, CFLAGS, CPPFLAGS and LDFLAGS are all now passed to
configure. CPPFLAGS is set for -fsanitize=address,undefined
to prevent unaligned access in HTSlib.

Adjusts warning options to work around false positives when
using undefined behaviour sanitizer, and to enable full format
checking on the gcc build with -Wextra.