Improve htslib#PRnum support for Cirrus-CI and GitHub Actions
Cirrus-CI stopped knowing the author's branch name at some point and
instead reported the branch name in the repository we're making a PR
against (pull/num).  So instead we added the ability to specify
htslib#NUM to the first line of a samtools PR to enable joint
samtools/htslib PRs to work.  This wasn't particularly robust and had
problems when not working as a PR (ie in the authors own repositories).

It was also completely absent from GitHub Actions.  Note there's no
way of knowing the PR number for GitHub actions short of doing a git
log, and it only has one commit in the log which is a merge commit.
I changed it to use pull/NUM/head instead of pull/NUM/merge, but
possibly it would be better done with "fetch-depth: 2" instead.  It's
no big deal though as we check the mergability elsewhere.

(cherry picked from commit eb0992ff8a99b895364a7a861b418a1beb77d540)