Fix various plot-bamstats issues in qual plots.
Fixes warning "warning: y2 axis range undefined or overflow, resetting to [0:0]"
which resulted in missing y-ticks.

Fixed quals-hm layout to be 2 distinct plots.

Adjusts the colour of the heatmap to be easier to read, avoiding
multiple dark zones in the colour wheel.

Fixes broken links in the HTML table for characters that need URI escaping.

Reduce width of plot so that "Number of bases" heatmap label is not clipped