View: add --save-counts option
Adds an option to store counts of records processed, accepted
and rejected by filtering to a file.  The data is saved as JSON,
but pretty-printed so it also can be processed fairly easily as a
simple text file.

Fixes #2038 (Using view command to filter the records from a bam
file, but also report the number of reads passed the filter or
dropped.)