Misc checksum updates
Added a --sanitize option, like fixmate and view, to correct broken
data prior to checksumming.  This can be useful for things that cannot
be supported in CRAM, such as storing MAPQ and CIGAR strings for
unmapped records or having READ1 and READ2 when not paired.

Added an extra column to act as a checksum of checksums, making
comparison easy.

Simplify the CRC aux code while also protecting against seq/qual "*".
A zero length string may mean the ks_string is never allocated so the
buffer is NULL.  This in turn resets the CRC value to 0 rather than
adding no data and repeating the input CRC.  This was what the aux tag
code was compensating for, but it's easier to just avoid the
possibility.

CIGAR is now MAPQ + CIGAR as the two fields logically belong together.