Add cigdup and cigarx options to samtools view -z sanitizer.
The first option collapses duplicate opcodes into a single one, so
2M3M becomes 5M.  This is enabled by default in the "on" and "all"
categories.

The second option replaces "=" and "X" with "M" (and collapsing as per
cigdup).  This is not technically an error and so is not enabled in
any mode unless explicitly asked for.  However it is a useful tool
when round-tripping to CRAM and back as it does not support the "="
and "X" opcodes.