Revamp the order-specific checksumming.
It's now done per readgroup, which means we add in the count at the
end rather than the start, permitting the same CRCs to be computed
once and added to multiple tallies.

This means we can now compute order-specific checksum on two files
worth of data (eg one read-group each) and "samtools merge" them.  The
order-specific checksums in the merged file still match those in their
original split files.