Canonicalise single-integer tags.
We see BAMs with degenerate tag enoding, such as using shorts for 0
values.  We also see BAMs using signed tags for positive values, which
if auto-generated (NM in CRAM) or round-tripped from SAM to BAM again
will get converted to unsigned.

Hence we have to canonicalise our tags before checksumming.

Note this still needs doing on byte array encodings, but it is unknown
if this is an issue in the wild yet.  Possibly they're better behaved.