Add cF to the tag exclusion list for "-a".
cF is an htslib specific private tag for storing CRAM meta-data.  It
indicates which of MD and/or NM are present when encoding using
embedded references, so the round-trip can be precise.  Htslib
transparently adds it and removes it during encoding and decoding,
however not all tools know about it so BAM files can be created with a
cF tag still present (eg io_lib's scramble or samtools pre 1.16).

Given it's a known difference, for ease of round-trip validation we
exclude this private tag.