Expand on qcfail and verbosity options.
We now split 3 way to all, qc-pass and qc-fail.  We no longer report
the qc fail/pass by default, but they can be added back with -q.

Rows that have zero counts are omitted, but can be shown with -v for
verbose mode (for ease of parsing).

We also now report read-group "-" for records not assigned to a
read-group.