Report read-groups in alphanumeric sorted order.
Also refactored the report printing code into its own function.

We could avoid a global if we built an array of char* pointers and
sorted those, but it would then need new hash table searches to find
them again which seems needless just for the sake of avoiding a static
global variable.