Fix samtools merge crash on BAM files with malformed headers
When the binary and text reference names disagree in a BAM header,
it's possible for the number of references apparently present to
increase when the header text is parsed.  This caused a crash when
samtools merge built its reference translation table because the
trans_tbl_t::tid_trans array was allocated in trans_tbl_init()
with the smaller size, and then trans_tbl_add_sq() used the bigger
one when trying to put data in it.

Fixed by moving the allocation to trans_tbl_add_sq(), so the
space allocated will always match the number of items iterated
over.

Thanks to @frostb1ten for the bug report.
Fixes #2127 (Heap Buffer Overflow in samtools merge within
trans_tbl_add_sq function)