Fix depth counts when overlap removal is enabled
The case where a match CIGAR operation ended exactly on the clip
point for overlap removal wasn't being handled correctly, causing
counts to be incorrectly incremented from the overlap clip point
to the end of the circular `hist` buffer.  Fix by adjusting the
existing test for fully clipped out operations to use `<=` instead
of `<`.

The same test is also adjusted for the DEL / REF_SLIP operators.
While it didn't cause a bug due to the way the counts were updated,
it does allow some unnecessary work to be skipped.

Put back an increment of `spos` for fully-clipped match operators,
which was removed in c93503ed.  Without it the wrong quality
values would be inspected for any later unclipped matches.

A test case for fully overlapping alignments is added to
mpileup/test/depth.reg.