Add a warning for samtools fastq on coordinate sorted data.
There can be legitimate cases of samtools fastq/fasta being run on
such data, for example when it is single ended, and we do not wish to
reorder the file. So making this a hard error may be problematic for
existing pipelines.

However it is also a cause of confusion and mistakes to users, so a
warning is necessary.

Fixes #2169