Don't use consensus N for "*" calls that are of insufficient depth.
The intention of the -d depth parameter is to only call a base when we
have sufficient confidence in the number of observations to call the
base type.  If it's insufficient depth then we call N.  However this
is an assertion that there is a base, we just don't wish to say which.

A column with mostly "*" (absent) and one single base (eg an overcall
error) should not be changed to N even when the depth is low as this
is asserting there is a base of some description there, which is
unlikely.

Fixes #2167