Fix mpileup -a bug
Caching the 2 references, and assuming if the first doesn't match then
we must be using the second, turns out to be incorrect.

It turns out our pileup code needs two (this ref and when using -a the
last reference to finish up the sequence), plus bam_mplp64_auto uses
two (current refence and potentially next reference if we're on the
last record of that reference), and these two aren't necessarily the
same pair.  So our revised number of in-flight references is now 3,
with better checking.

Updated the test data to trigger the bug (under address sanitizer), to
verify the fix works before the code changes were added.

Fixes #2227