Speed up samtools consensus by 20-25% for long-read data.
The defaule Bayesian method does a lot of local modification of
mapping quality via the local mismatch count.  This is floating point
based, but the variable being integer means lots of unnecessary type
conversions happen.  It's surprising this is such a significant
performance cost compared to everything else that goes on (including a
division which can be avoided using lookup tables, but isn't
significant).