Add a link to samtools.1 manpage for global opts in usage.
Addresses issue raised in #2236 with a pointer to where to find the
list of acceptable input-fmt-option and output-fmt-option keywords.
Many people don't think to look at the global samtools.1 man-page for
such things.  Arguably we should edit every manpage too referring back
to this, but this is an easy and more discoverable fix.